+++ ci/test_suites/authenticators_gcp/deploy_function.sh
@@ -68,7 +68,7 @@ deploy_function() {
   cd "$GCF_SOURCE_DIR"
 
   echo "-- Deploying function: $GCF_FUNC_NAME"
-  gcloud functions deploy "$GCF_FUNC_NAME" --runtime python37 --trigger-http --quiet
+  gcloud functions deploy "$GCF_FUNC_NAME" --runtime python37 --trigger-http --quiet -no-gen2
 
   echo '-> deploy_function done'
 }