+++ lib/conjur/conjur_config.rb
@@ -44,6 +44,10 @@ module Conjur
       events_to_sns_interval: 30, # Events to SNS interval should be every 30 seconds
       tenant_id: @tenant_id,
       tenant_name: @tenant_name,
+      # keep the value below to allow testing local....
+      # conjur_pubsub_enabled: true,
+      # conjur_pubsub_iam_role: 'arn:aws:iam::123456789012:role/conjur-pubsub-role',
+      # conjur_pubsub_sns_topic: "arn:aws:sns:us-east-1:000000000000:MyDeveloperTopic.fifo",
       tenant_env: @tenant_env,
       tenant_region: @tenant_region,
       policy_factories_path: 'conjur/factories'