+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Nothing should go in this section, please add to the latest unreleased version
   (and update the corresponding date), or add a new version.
 
+## [1.0.40-cloud] - 2024-09-17
+### Fixed
+- Fixed assume role when credentials expired
+
 ## [1.0.39-cloud] - 2024-09-10
 ### Added
 - changing status controller to work without transaction